# public/api.yaml
openapi: 3.0.0
paths:
  /api/loginVisitor:
    post:
      tags:
        - Adherent
      summary: 'Logs in a visitor'
      description: "À visitor's connection request."
      operationId: fad7ad7399b7d621b5eca795403beda4
      requestBody:
        description: "Visitor's email and password"
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              type: object
      responses:
        '200':
          description: 'Successful login'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: success }
                  user: { $ref: '#/components/schemas/User' }
                  authorisation: { properties: { token: { type: string, example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9... }, type: { type: string, example: bearer } }, type: object }
                type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: Unauthorized }
                type: object
  /api/registerVisitor:
    post:
      tags:
        - Adherent
      summary: 'Logs in a visitor'
      description: 'The request for registration.'
      operationId: b1f737f531e830162dfb399b3242aa7f
      requestBody:
        description: "Visitor's register"
        required: true
        content:
          application/json:
            schema:
              required:
                - login
                - nom
                - email
                - prenom
                - pseudo
              properties:
                email:
                  type: string
                  format: email
                login:
                  type: string
                nom:
                  type: string
                prenom:
                  type: string
                pseudo:
                  type: string
              type: object
      responses:
        '200':
          description: 'Successful register'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: success }
                  user: { $ref: '#/components/schemas/User' }
                  authorisation: { properties: { token: { type: string, example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9... }, type: { type: string, example: bearer } }, type: object }
                type: object
        '422':
          description: Unauthorized
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: Unauthorized }
                type: object
  '/api/adherent/{id}/avatar':
    post:
      tags:
        - Adherent
      summary: 'Logs in a visitor'
      description: 'Update the avatar in storage.'
      operationId: b57763c8a839e66bab28580313a481a7
      requestBody:
        description: 'Update avatar'
        required: true
        content:
          application/json:
            schema:
              required:
                - avatar
              properties:
                avatar:
                  type: string
              type: object
      responses:
        '200':
          description: 'Successful update avatar'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: success }
                  user: { $ref: '#/components/schemas/User' }
                  authorisation: { type: object }
                type: object
        '422':
          description: Unauthorized
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: Unauthorized }
                type: object
...